+++ apps/passport-client/components/screens/ScannedTicketScreens/SecondPartyTicketVerifyScreen.tsx
